+++ src/measure.h
@@ -18,12 +18,12 @@ void measure_start(struct measure_t *data);
 void measure_stop(struct measure_t *data);
 
 // Measure getters: all returned values are in microseconds
-uint32_t measure_get_min_cycles(struct measure_t *data);
-uint32_t measure_get_max_cycles(struct measure_t *data);
-uint32_t measure_get_avg_cycles(struct measure_t *data);
-uint32_t measure_get_min_nsec(struct measure_t *data);
-uint32_t measure_get_max_nsec(struct measure_t *data);
-uint32_t measure_get_avg_nsec(struct measure_t *data);
+uint64_t measure_get_min_cycles(struct measure_t *data);
+uint64_t measure_get_max_cycles(struct measure_t *data);
+uint64_t measure_get_avg_cycles(struct measure_t *data);
+uint64_t measure_get_min_nsec(struct measure_t *data);
+uint64_t measure_get_max_nsec(struct measure_t *data);
+uint64_t measure_get_avg_nsec(struct measure_t *data);
 uint64_t measure_get_count(struct measure_t *data);
 
 #endif // MEASURE_H_
